+++ neo/private_key_test.go
@@ -31,11 +31,12 @@ func TestPrivateKey(t *testing.T) {
 		})
 	})
 
-	t.Run("NewWIFFromPrivateKey()", func(t *testing.T) {
+	t.Run("WIF()", func(t *testing.T) {
 		t.Run("HappyCase", func(t *testing.T) {
-			p, _ := neo.NewPrivateKey()
-			wif, err := p.NewWIFFromPrivateKey()
-			assert.Equal(t, 52, len(*wif))
+			p, err := neo.NewPrivateKey()
+			assert.NoError(t, err)
+			wif, err := p.WIF()
+			assert.Equal(t, 52, len(wif))
 			assert.NoError(t, err)
 		})
 	})